requirements.txt
aiogram==2.25.1
aiohttp
python-dotenv